Drag and drop corners is ok, we still have an inversion and corners do not match visually but it's almost done.

+++ veDDDa/veDDDa/MainWindow.xaml.cs
@@ -173,24 +173,24 @@ namespace veDDDa
             GL.Uniform1(eyePosLoc, (Eye == EEye.LEFT ? -1.0f : 1.0f));
 
             var topLeftX = GL.GetUniformLocation(_program, "TopLeftX");
-            GL.Uniform1(topLeftX, _model.TopLeft.Left);
+            GL.Uniform1(topLeftX, (float)_model.TopLeft.Left);
             var topLeftY = GL.GetUniformLocation(_program, "TopLeftY");
-            GL.Uniform1(topLeftY, _model.TopLeft.Top);
+            GL.Uniform1(topLeftY, (float)_model.TopLeft.Top);
 
             var topRightX = GL.GetUniformLocation(_program, "TopRightX");
-            GL.Uniform1(topRightX, _model.TopRight.Left);
+            GL.Uniform1(topRightX, (float)_model.TopRight.Left);
             var topRightY = GL.GetUniformLocation(_program, "TopRightY");
             GL.Uniform1(topRightY, _model.TopRight.Top);
 
             var bottomLeftX = GL.GetUniformLocation(_program, "BottomLeftX");
-            GL.Uniform1(bottomLeftX, _model.BottomLeft.Left);
+            GL.Uniform1(bottomLeftX, (float)_model.BottomLeft.Left);
             var bottomLeftY = GL.GetUniformLocation(_program, "BottomLeftY");
-            GL.Uniform1(bottomLeftY, _model.BottomLeft.Top);
+            GL.Uniform1(bottomLeftY, (float)_model.BottomLeft.Top);
 
             var bottomRightX = GL.GetUniformLocation(_program, "BottomRightX");
-            GL.Uniform1(bottomRightX, _model.BottomRight.Left);
+            GL.Uniform1(bottomRightX, (float)_model.BottomRight.Left);
             var bottomRightY = GL.GetUniformLocation(_program, "BottomRightY");
-            GL.Uniform1(bottomRightY, _model.BottomRight.Top);
+            GL.Uniform1(bottomRightY, (float)_model.BottomRight.Top);
 
 
         }
